Bugfix: Add to existing file
If WriteGenericNc is invoked with a filename that already exists, the
variable is now added to the existing file.
The function checks for already existing dimensions that might be the
same as in the new variable. If that’s the case, the variable is
attributed to the existing dimensions. If not, new dimensions are added
to the file.