do_anomaly option in GetWaves
GetWaves has now an option whether or not you want to FFT from
anomalies or keeping the mean in.
Normally, you don’t need the anomalies, as the mean goes into wave-0,
and the anomalies into higher order waves.
Sometimes it’s simply nice to have the option to create the anomalies
first, to be sure.