Vstar, Wstar
Changed input of ComputeVstar, data, from list to dictionary. It is
still possible to pass a file name, but if data is passed instead of a
file name, the data has to be a dictionary with the field names defined
by temp, vcomp, pfull.

Added ComputeWstar, which has the same structure as ComputeVstar, but
computes omega* from omega, v, T, p, phi.