Added eof,ComputeAnnularMode choices
- It is now possible to actively choose which hemisphere to consider
for the Annular Mode computations. Setting `hemi` to 'NH' or 'SH'
will concentrate on the northern or southern hemispheres respectively.
The default, 'infer', will decide based on the input latitudes:
`if mean(lat) >= 0`, then the NH will be considered, otherwise the
southern hemisphere.

- It is also possible to choose the detrending method used to compute
the EOF: Either 'linear' or 'constant'. This goes directly into the
call of `scipy.signal.detrend`. See documentation for that function
for more information.